Add HTML structure, CSS template for tabs
The idea behind this site is to eventually use a tab system to have the
page transition between different content with DOM manipulation instead
of having separate links. The HTML/CSS structure is designed with that
in mind, making use of a simple content div that will eventually be used
to display the various pages, and CSS that can enforce some consistency
between the different tabs. Only thing I want to square away before
beginning to script this is the button styling.